edit: seperate attachment class
