+++ src/services/schoolNotice.ts
@@ -1,24 +1,5 @@
 import { Injectable } from '@decorators/di';
-
-export interface IAttachment {
-    filename: string; // ex) test.jpg
-    data: Buffer;
-    mimetype: string; // ex) image/jpeg (more info: https://developer.mozilla.org/ko/docs/Web/HTTP/Basics_of_HTTP/MIME_types)
-}
-
-export class Attachment implements IAttachment {
-    filename: string;
-    data: Buffer;
-    mimetype: string;
-    constructor(filename: string, data: Buffer, mimetype: string) {
-        this.filename = filename;
-        this.data = data;
-        this.mimetype = mimetype;
-    }
-    static fromObject(data: IAttachment) {
-        return new Attachment(data.filename, data.data, data.mimetype);
-    }
-}
+import { Attachment, IAttachment } from '../types/attachment';
 
 export interface ISchoolNotice {
     id: number;
